AgenticAI/ResearchWriter/agent-app/pyproject.toml
[project]
name = "agent-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=5.45.0",
    "openai-agents>=0.3.0",
    "python-dotenv>=1.1.1",
    "sendgrid>=6.12.4",
]
